修复创建角色时的异常

+++ MMORPG/Assets/Scripts/Event/JoinedMapEvent.cs
@@ -3,9 +3,9 @@ namespace MMORPG.Event
     public class JoinedMapEvent
     {
         public int MapId { get; }
-        public int CharacterId { get; }
+        public long CharacterId { get; }
 
-        public JoinedMapEvent(int mapId, int characterId)
+        public JoinedMapEvent(int mapId, long characterId)
         {
             MapId = mapId;
             CharacterId = characterId;
